cleaned comments and split graphql file
