+++ src/scenes/Root/CommentsBar/CommentItem/CommentItem.graphql
@@ -1,0 +1,5 @@
+mutation DeleteComment($id: ID!) {
+  deleteComment(id: $id) {
+    __typename
+  }
+}
